django_versions/sbitter_postgres_site/requirements.txt
Django==1.4.5
South==0.7.6
argparse==1.2.1
django-tagging==0.3.1
wsgiref==0.1.2
